backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dateutil==2.9.0.post0
pytest==8.3.4
httpx==0.28.1
ruff==0.9.6
